move new coupon helper functions into client/lib/upgrades/actions/cart.js
